Make exit a long, so comparison works for all syscalls.
git-svn-id: http://audit-test.svn.sourceforge.net/svnroot/audit-test/trunk/tests@924 df539fd5-5ffc-0310-83f5-f045c3a6a7d5

+++ audit/include/utils.h
@@ -83,7 +83,7 @@ void context_setpid(struct audit_data *, int);
 void context_setexperror(struct audit_data *, int);
 void context_setbegin(struct audit_data *);
 void context_setend(struct audit_data *);
-void context_setresult(struct audit_data *, int, int);
+void context_setresult(struct audit_data *, long, int);
 void context_setipc(struct audit_data *, int, int, int, long);
 void context_setsockaddr(struct audit_data *, unsigned char *, size_t);
 int context_setcwd(struct audit_data *);
